requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
alembic==1.12.1
authlib==1.2.1
aiosmtplib==3.0.1
bcrypt==4.0.1
certifi==2025.7.14
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.116.1
frozenlist==1.7.0
faiss-cpu
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-experimental
langchain-mongodb==0.6.2
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.8
lark==1.2.2
marshmallow==3.26.1
motor==3.7.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.97.1
orjson==3.11.1
packaging==25.0
passlib==1.7.4
pdf2image==1.17.0
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydash==8.0.5
pymongo==4.13.2
pytesseract==0.3.13
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
psycopg2-binary
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
stripe==7.0.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0


